+++ NEWS.md
@@ -1,12 +1,21 @@
 ![https://www.youtube.com/watch?v=TaR-bwZxfuA](https://r2symbols.obi.obianom.com/ezgif-3-efa9cce31d.gif)
 
+# r2symbols 1.5
+## Planned Features
+ - [ ] User interface for easy use of symbols
+
+
+# r2symbols 1.4
+## Features
+ - [ ] Fetch multiple symbols by keyword
+ - [ ] Fetch multiple symbols by category
+ 
 # r2symbols 1.3
 ## Features
 
  - [x] Addition of over 800 new symbols
  - [x] Fix of bugs associated with changing symbols settings in Rmarkdown
- - [ ] User interface for easy use of symbols
- 
+
  
 # r2symbols 1.2